Move view room dispatches into the Room View Store to tidy up
